pyproject.toml
[tool.poetry]
name = "cied-app"
version = "0.1.0"
description = ""
authors = ["YADHU KRISHNA P V <yadhu2815@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
